+++ src/Products/PloneMeeting/utils.py
@@ -594,7 +594,7 @@ def sendMailIfRelevant(obj,
         return
     # Do not send mail if the (not custom) event is unknown.
     if not customEvent and event not in cfg.getMailItemEvents() and \
-       event not in cfg.getMailMeetingEvents():
+            event not in cfg.getMailMeetingEvents():
         return
     # Ok, send a mail. Who are the recipients ?
     recipients = []
@@ -661,6 +661,7 @@ def sendMailIfRelevant(obj,
             unique_emails.append(email)
             unique_email_recipients.append(recipient)
         mail_subject, mail_body = sendMail(unique_email_recipients, obj, event, mapping=mapping)
+        logger.info("Mail(s) sent to :" +  " ".join(recipients))
     debug = debug or obj.REQUEST.get('debug_sendMailIfRelevant', False)
     if debug:
         return recipients, mail_subject, mail_body